# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Todo App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /todos:
    x-swagger-router-controller: todo
    get:
      description: Return All Todos
      operationId: getTodos
      parameters:
        - name: all
          in: query
          description: If all is false show only not completed list
          required: false
          type: boolean
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetTodosListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /todo:
    x-swagger-router-controller: todo
    post:
      description: Save a Todo Item
      operationId: saveTodo
      parameters:
        - name: todo
          in: body
          description: Todo Body
          required: true
          schema:
            $ref: "#/definitions/Todo"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /todo/{id}:
    x-swagger-router-controller: todo
    get:
      description: Get a Todo with Id
      operationId: getTodo
      parameters:
        - name: id
          type: string
          in: path
          description: Todo Id
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetTodoResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Update a Todo with Id
      operationId: updateTodo
      parameters:
        - name: id
          type: string
          in: path
          description: Todo Id
          required: true
        - name: todo
          in: body
          description: Todo Object
          required: true
          schema:
            $ref: "#/definitions/Todo"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: delete a Todo
      # define the parameters
      operationId: deleteTodo
      parameters:
        - name: id
          description: Todo id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  Todo:
    type: object
    properties:
      text:
        type: string
        description: Description of the Todo Item
      completed:
        type: boolean
        description: If the todo item was complete
      completedAt:
        type: number
        description: Time when the Todo was completed
    required:
      - text
  GetTodosListResponse:
    required:
      - todos
    properties:
      todos:
        type: array
        items:
          type: object
          properties:
            _id:
              type: string
            text:
              type: string
            completed:
              type: boolean
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment 
    required:
      - success
      - description
  GetTodoResponse:
    required:
      - _id
      - text
      - completed
    properties:
      _id:
        type: string
      text: 
        type: string
      completed:
        type: boolean